feat(#516): forcing git initialisation by a new flag `--force-git-init`
- refactor git related tests into an own integration test module
